Read CSV file for modelling a NN VIA NNTOOL
As shown in: https://youtu.be/icNFe2a8s-c

Brief tutorial that creates a neural network using the nntool command. This allows access to the the Network/Data Manager window, which, in turn facilitates importing, creating, using, and exporting neural networks and data.

The case of study are sensor readings from a robot that represent distances (left and front left) with their corresponding outputs (left and right motor speeds). The implementation goes from CSV file reading to NN configuration and performance.